add start and stop device,
the callback works but annoyingly there is no sound.
That is because Once the sample rate is set, if I recheck the sample rate I realised that it is not the current sample rate of the device.
This is strange as the device sent no error when I wetted the property, what happens ?